bunch of espn stuff

+++ src/main/java/ncollins/model/espn/Status.java
@@ -1,12 +1,12 @@
 package ncollins.model.espn;
 
 public class Status {
-    private float activatedDate;
-    private float createdAsLeagueType;
-    private float currentLeagueType;
-    private float currentMatchupPeriod;
-    private float finalScoringPeriod;
-    private float firstScoringPeriod;
+    private long    activatedDate;
+    private int     createdAsLeagueType;
+    private int     currentLeagueType;
+    private int     currentMatchupPeriod;
+    private int     finalScoringPeriod;
+    private int     firstScoringPeriod;
     private boolean isActive;
     private boolean isExpired;
     private boolean isFull;
@@ -14,19 +14,19 @@ public class Status {
     private boolean isToBeDeleted;
     private boolean isViewable;
     private boolean isWaiverOrderEdited;
-    private float latestScoringPeriod;
-    private float standingsUpdateDate;
-    private float teamsJoined;
-    private float transactionScoringPeriod;
-    private float waiverLastExecutionDate;
+    private int     latestScoringPeriod;
+    private long    standingsUpdateDate;
+    private int     teamsJoined;
+    private int     transactionScoringPeriod;
+    private long    waiverLastExecutionDate;
 
     // GET
-    public float getActivatedDate() { return activatedDate; }
-    public float getCreatedAsLeagueType() { return createdAsLeagueType; }
-    public float getCurrentLeagueType() { return currentLeagueType; }
-    public float getCurrentMatchupPeriod() { return currentMatchupPeriod; }
-    public float getFinalScoringPeriod() { return finalScoringPeriod; }
-    public float getFirstScoringPeriod() { return firstScoringPeriod; }
+    public long getActivatedDate() { return activatedDate; }
+    public int getCreatedAsLeagueType() { return createdAsLeagueType; }
+    public int getCurrentLeagueType() { return currentLeagueType; }
+    public int getCurrentMatchupPeriod() { return currentMatchupPeriod; }
+    public int getFinalScoringPeriod() { return finalScoringPeriod; }
+    public int getFirstScoringPeriod() { return firstScoringPeriod; }
     public boolean getIsActive() { return isActive; }
     public boolean getIsExpired() { return isExpired; }
     public boolean getIsFull() { return isFull; }
@@ -34,19 +34,19 @@ public class Status {
     public boolean getIsToBeDeleted() { return isToBeDeleted; }
     public boolean getIsViewable() { return isViewable; }
     public boolean getIsWaiverOrderEdited() { return isWaiverOrderEdited; }
-    public float getLatestScoringPeriod() { return latestScoringPeriod; }
-    public float getStandingsUpdateDate() { return standingsUpdateDate; }
-    public float getTeamsJoined() { return teamsJoined; }
-    public float getTransactionScoringPeriod() { return transactionScoringPeriod; }
-    public float getWaiverLastExecutionDate() { return waiverLastExecutionDate; }
+    public int getLatestScoringPeriod() { return latestScoringPeriod; }
+    public long getStandingsUpdateDate() { return standingsUpdateDate; }
+    public int getTeamsJoined() { return teamsJoined; }
+    public int getTransactionScoringPeriod() { return transactionScoringPeriod; }
+    public long getWaiverLastExecutionDate() { return waiverLastExecutionDate; }
 
     // SET
-    public void setActivatedDate(float activatedDate) { this.activatedDate = activatedDate; }
-    public void setCreatedAsLeagueType(float createdAsLeagueType) { this.createdAsLeagueType = createdAsLeagueType; }
-    public void setCurrentLeagueType(float currentLeagueType) { this.currentLeagueType = currentLeagueType; }
-    public void setCurrentMatchupPeriod(float currentMatchupPeriod) { this.currentMatchupPeriod = currentMatchupPeriod; }
-    public void setFinalScoringPeriod(float finalScoringPeriod) { this.finalScoringPeriod = finalScoringPeriod; }
-    public void setFirstScoringPeriod(float firstScoringPeriod) { this.firstScoringPeriod = firstScoringPeriod; }
+    public void setActivatedDate(long activatedDate) { this.activatedDate = activatedDate; }
+    public void setCreatedAsLeagueType(int createdAsLeagueType) { this.createdAsLeagueType = createdAsLeagueType; }
+    public void setCurrentLeagueType(int currentLeagueType) { this.currentLeagueType = currentLeagueType; }
+    public void setCurrentMatchupPeriod(int currentMatchupPeriod) { this.currentMatchupPeriod = currentMatchupPeriod; }
+    public void setFinalScoringPeriod(int finalScoringPeriod) { this.finalScoringPeriod = finalScoringPeriod; }
+    public void setFirstScoringPeriod(int firstScoringPeriod) { this.firstScoringPeriod = firstScoringPeriod; }
     public void setIsActive(boolean isActive) { this.isActive = isActive; }
     public void setIsExpired(boolean isExpired) { this.isExpired = isExpired; }
     public void setIsFull(boolean isFull) { this.isFull = isFull; }
@@ -54,9 +54,9 @@ public class Status {
     public void setIsToBeDeleted(boolean isToBeDeleted) { this.isToBeDeleted = isToBeDeleted; }
     public void setIsViewable(boolean isViewable) { this.isViewable = isViewable; }
     public void setIsWaiverOrderEdited(boolean isWaiverOrderEdited) { this.isWaiverOrderEdited = isWaiverOrderEdited; }
-    public void setLatestScoringPeriod(float latestScoringPeriod) { this.latestScoringPeriod = latestScoringPeriod; }
-    public void setStandingsUpdateDate(float standingsUpdateDate) { this.standingsUpdateDate = standingsUpdateDate; }
-    public void setTeamsJoined(float teamsJoined) { this.teamsJoined = teamsJoined; }
-    public void setTransactionScoringPeriod(float transactionScoringPeriod) { this.transactionScoringPeriod = transactionScoringPeriod; }
-    public void setWaiverLastExecutionDate(float waiverLastExecutionDate) { this.waiverLastExecutionDate = waiverLastExecutionDate; }
+    public void setLatestScoringPeriod(int latestScoringPeriod) { this.latestScoringPeriod = latestScoringPeriod; }
+    public void setStandingsUpdateDate(long standingsUpdateDate) { this.standingsUpdateDate = standingsUpdateDate; }
+    public void setTeamsJoined(int teamsJoined) { this.teamsJoined = teamsJoined; }
+    public void setTransactionScoringPeriod(int transactionScoringPeriod) { this.transactionScoringPeriod = transactionScoringPeriod; }
+    public void setWaiverLastExecutionDate(long waiverLastExecutionDate) { this.waiverLastExecutionDate = waiverLastExecutionDate; }
 }
